+++ index.html
@@ -57,6 +57,6 @@
         <div class="wall-left"></div>
         <div class="wall-right"></div>
     </div>
-    <script src="main.js"></script>
+    <script src="main.js" type="module" mimetypes="application/javascript"></script>
 </body>
 </html>
